// src/views/platform/company/list/utils/types.ts
interface FormItemProps {
  /** 用于判断是`新增`还是`修改` */
  title: string;
  /** 公司id */
  id?: string;
  /** 公司简称 */
  abbr: string;
  /** 公司名 */
  name: string;
  /** 公司社会统一信用代码 */
  uscc: string;
  /** 联系人 */
  contactName: string;
  /** 联系电话 */
  contactPhone: string;
  /** 邮箱 */
  email: string;
  /** 通信地址 */
  address: string;
  /** 绑定域名 */
  website: string;
  /** 账号数量 */
  accountCount: number;
  /** 租户套餐 */
  packageId: number;
  /** 备注 */
  remark: string;
  /** 账号 */
  username: string;
  /** 密码 */
  password: string;
}

interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };